Add maps to components
